Add files via upload
These Excel files are used to generate input data to vensim files. OBS there is no "ver" history of these. Date 2023-08-09. I avoid changing any names etc now due to that the vensim files are uploaded and that info in the models and Excel files needs to be updated to function if this is implemented. //Gary